src/requirements.txt
attrs==22.1.0
jsonschema==4.9.1
pyrsistent==0.18.1
PyYAML==6.0
argparse==1.4.0